♻️ Refactor _compat, prepare compatibility for Pydantic v1 in v2
