+++ fastapi/_compat/__init__.py
@@ -1,0 +1,49 @@
+from .main import BaseConfig as BaseConfig
+from .main import PydanticSchemaGenerationError as PydanticSchemaGenerationError
+from .main import RequiredParam as RequiredParam
+from .main import Undefined as Undefined
+from .main import UndefinedType as UndefinedType
+from .main import Url as Url
+from .main import Validator as Validator
+from .main import _get_model_config as _get_model_config
+from .main import _is_error_wrapper as _is_error_wrapper
+from .main import _is_model_class as _is_model_class
+from .main import _is_model_field as _is_model_field
+from .main import _is_undefined as _is_undefined
+from .main import _model_dump as _model_dump
+from .main import _model_rebuild as _model_rebuild
+from .main import copy_field_info as copy_field_info
+from .main import create_body_model as create_body_model
+from .main import evaluate_forwardref as evaluate_forwardref
+from .main import get_annotation_from_field_info as get_annotation_from_field_info
+from .main import get_cached_model_fields as get_cached_model_fields
+from .main import get_compat_model_name_map as get_compat_model_name_map
+from .main import get_definitions as get_definitions
+from .main import get_missing_field_error as get_missing_field_error
+from .main import get_schema_from_model_field as get_schema_from_model_field
+from .main import is_bytes_field as is_bytes_field
+from .main import is_bytes_sequence_field as is_bytes_sequence_field
+from .main import is_scalar_field as is_scalar_field
+from .main import is_scalar_sequence_field as is_scalar_sequence_field
+from .main import is_sequence_field as is_sequence_field
+from .main import serialize_sequence_value as serialize_sequence_value
+from .main import (
+    with_info_plain_validator_function as with_info_plain_validator_function,
+)
+from .model_field import ModelField as ModelField
+from .shared import PYDANTIC_V2 as PYDANTIC_V2
+from .shared import PYDANTIC_VERSION_MINOR_TUPLE as PYDANTIC_VERSION_MINOR_TUPLE
+from .shared import field_annotation_is_scalar as field_annotation_is_scalar
+from .shared import (
+    is_uploadfile_or_nonable_uploadfile_annotation as is_uploadfile_or_nonable_uploadfile_annotation,
+)
+from .shared import (
+    is_uploadfile_sequence_annotation as is_uploadfile_sequence_annotation,
+)
+from .shared import lenient_issubclass as lenient_issubclass
+from .shared import sequence_types as sequence_types
+from .shared import value_is_sequence as value_is_sequence
+from .v1 import CoreSchema as CoreSchema
+from .v1 import GetJsonSchemaHandler as GetJsonSchemaHandler
+from .v1 import JsonSchemaValue as JsonSchemaValue
+from .v1 import _normalize_errors as _normalize_errors
